% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_affordable_housing}
\alias{oreg_affordable_housing}
\title{Affordable housing inventory}
\usage{
oreg_affordable_housing(city = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{city}{Filter by city}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Affordable housing inventory
}
