% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_agencies}
\alias{oreg_agencies}
\title{Oregon agencies, boards, and commissions}
\usage{
oreg_agencies(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Oregon agencies, boards, and commissions
}
