% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_businesses_county}
\alias{oreg_businesses_county}
\title{Active businesses by county}
\usage{
oreg_businesses_county(county = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{county}{Filter by county name}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Active businesses by county
}
