% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_campaign_penalties}
\alias{oreg_campaign_penalties}
\title{Campaign finance penalty notices}
\usage{
oreg_campaign_penalties(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Campaign finance penalty notices
}
