% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_cc_courses}
\alias{oreg_cc_courses}
\title{Community college course listing}
\usage{
oreg_cc_courses(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Community college course listing
}
