% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_ccb_licenses}
\alias{oreg_ccb_licenses}
\title{CCB (Construction Contractors Board) active licenses}
\usage{
oreg_ccb_licenses(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
CCB (Construction Contractors Board) active licenses
}
