% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_contractor_licenses}
\alias{oreg_contractor_licenses}
\title{Building Codes Division - Active contractor/individual licenses}
\usage{
oreg_contractor_licenses(business_name = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{business_name}{Filter by name}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Building Codes Division - Active contractor/individual licenses
}
