% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_data_inventory}
\alias{oreg_data_inventory}
\title{State agency data inventory}
\usage{
oreg_data_inventory(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
State agency data inventory
}
