% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_disease_data}
\alias{oreg_disease_data}
\title{Weekly reportable disease data}
\usage{
oreg_disease_data(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Weekly reportable disease data
}
