% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_emergency_managers}
\alias{oreg_emergency_managers}
\title{OEM emergency managers (local and tribal)}
\usage{
oreg_emergency_managers(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
OEM emergency managers (local and tribal)
}
