% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_farm_products}
\alias{oreg_farm_products}
\title{Farm products master list}
\usage{
oreg_farm_products(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Farm products master list
}
