% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_fertilizer_stops}
\alias{oreg_fertilizer_stops}
\title{Fertilizer program stop sales}
\usage{
oreg_fertilizer_stops(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Fertilizer program stop sales
}
