% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_hazmat_sites}
\alias{oreg_hazmat_sites}
\title{Hazardous materials storage sites}
\usage{
oreg_hazmat_sites(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Hazardous materials storage sites
}
