% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_housing_grants}
\alias{oreg_housing_grants}
\title{OHCS housing grant opportunities}
\usage{
oreg_housing_grants(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
OHCS housing grant opportunities
}
