% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_libraries}
\alias{oreg_libraries}
\title{Oregon library directory}
\usage{
oreg_libraries(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Oregon library directory
}
