% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_medicaid_funding}
\alias{oreg_medicaid_funding}
\title{Oregon Medicaid prioritized list funding line}
\usage{
oreg_medicaid_funding(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Oregon Medicaid prioritized list funding line
}
