% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_medical_licenses}
\alias{oreg_medical_licenses}
\title{Oregon Medical Board license type and status counts}
\usage{
oreg_medical_licenses(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Oregon Medical Board license type and status counts
}
