% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_nonprofits}
\alias{oreg_nonprofits}
\title{Active nonprofit corporations}
\usage{
oreg_nonprofits(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Active nonprofit corporations
}
