% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_notaries}
\alias{oreg_notaries}
\title{Notary public directory}
\usage{
oreg_notaries(last_name = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{last_name}{Filter by last name}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Notary public directory
}
