% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_rental_housing_partners}
\alias{oreg_rental_housing_partners}
\title{Affordable rental housing partner database}
\usage{
oreg_rental_housing_partners(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Affordable rental housing partner database
}
