% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_sos_payroll}
\alias{oreg_sos_payroll}
\title{Secretary of State payroll (current biennium)}
\usage{
oreg_sos_payroll(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Secretary of State payroll (current biennium)
}
