% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_trademarks}
\alias{oreg_trademarks}
\title{Active trademark registrations}
\usage{
oreg_trademarks(q = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{q}{Search trademark descriptions}

\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Active trademark registrations
}
