% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_training_providers}
\alias{oreg_training_providers}
\title{Eligible training provider list}
\usage{
oreg_training_providers(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Eligible training provider list
}
