% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_ucc_filings}
\alias{oreg_ucc_filings}
\title{UCC (Uniform Commercial Code) filings entered last month}
\usage{
oreg_ucc_filings(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
UCC (Uniform Commercial Code) filings entered last month
}
