% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_view}
\alias{oreg_view}
\title{Query any Oregon dataset by view ID (generic SODA access)}
\usage{
oreg_view(view_id, select = NULL, where = NULL,
                      order = NULL, q = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{view_id}{Four-by-four Socrata view ID}

\item{select}{SoQL select clause}

\item{where}{SoQL where clause}

\item{order}{SoQL order clause}

\item{q}{Full-text search within dataset}

\item{limit}{Max rows (default 1000)}

\item{offset}{Row offset}

}
\value{
tibble
}
\description{
Query any Oregon dataset by view ID (generic SODA access)
}
