% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_voter_registration}
\alias{oreg_voter_registration}
\title{Voter registration data}
\usage{
oreg_voter_registration(county = NULL, party = NULL,
                                    where = NULL, limit = 1000)
}
\arguments{
\item{county}{Filter by county name}

\item{party}{Filter by party name}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Registration counts by county, party, and district.
}
