% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_voting_districts}
\alias{oreg_voting_districts}
\title{Voting districts by precinct}
\usage{
oreg_voting_districts(county = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{county}{Filter by county}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Voting districts by precinct
}
