% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_warn_notices}
\alias{oreg_warn_notices}
\title{WARN (Worker Adjustment and Retraining Notification) notices}
\usage{
oreg_warn_notices(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
WARN (Worker Adjustment and Retraining Notification) notices
}
