% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_workers_comp_claims}
\alias{oreg_workers_comp_claims}
\title{Workers' compensation record level claims}
\usage{
oreg_workers_comp_claims(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Workers' compensation record level claims
}
