% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oregon.R
\name{oreg_workers_comp_employers}
\alias{oreg_workers_comp_employers}
\title{Active workers' compensation employer database}
\usage{
oreg_workers_comp_employers(business_name = NULL, county_code = NULL,
                                        where = NULL, limit = 1000)
}
\arguments{
\item{business_name}{Filter by employer name}

\item{county_code}{Filter by county code}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Active workers' compensation employer database
}
